Display episodes from the API in Index.xml.js
Use uitzendinggemist.js to load episodes to show in the episode index.
The code in `client/templates/Index.xml.js` should be refactored, as
selecting the nicest image for the episode and removing `&`s should be
done elsewhere.

Also, it shouldn't be necessary to pass the `baseURL` to
`episodeLockups()` or to have `this.BASEURL` in the template, as it's
not variable. It's probably `client/js/Presenter.js`'s responsibility.