Remove the current episode from the series episodes list
`Episode.find` uses `UitzendingGemist.Series.find` to first get the
`Series`, which holds a list of `Episodes`. The `Episode` returned by
`Episode.find` comes from that list.

The `Episode.series` is the `Series` from
`UitzendingGemist.Series.find`, including the episode list. Now, the
`Episode` is removed from that list before creating the `Series`, so the
`Episode` isn't listed in the shelf on the episode view.