Add Episode.date and Episode.broadcasted_at
Get the broadcasted_at date from the UitzendingGemist API and add a
`Episode.broadcasted_at`, which returns a humanized date string (like
14/11/2015 17:58).

It's used in the subtitles for the "more"-items and to the top row on
the episode view. It could be a good idea to add it to the subtitles on
the recent and popular views, but since that doesn't fit in most of the
cases, the subtitles (and titles) should first scroll across like they
do in iTunes Movies.

Closes #7.