Catch JSON parser errors in uitzendinggemist.js
Instead of crashing, JSON parser errors now show the error in an alert,
which is unfriendly, but keeps the app running instead of making the
whole thing unusable.

The messages in the alerts could use some TLC, as they don't display
filenames, line numbers, or even the body that couldn't be parsed,
making it almost impossible to use the alert message to debug.