Check if `data.series` exists before calling `data.series.image`
It can take a while for new episodes to get episode images, so
Episode.js checks the episode stills, the episode image, and the series
image in that order. It didn't check if there was a series before
attempting to get the series image, which caused errors for nearly all
episodes in the first half hour after release or so.